requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.2.16
langchain-google-genai==0.0.6
faiss-cpu==1.8.0
sentence-transformers==2.2.2
